split app.vue to adminapp.vue and userapp.vue. split navbar into adminnavbar and usernavbar
