+++ src/router/index.js
@@ -11,7 +11,7 @@ import AddCrewMember from '@/components/CrewMember/AddCrewMember.vue'
 import FindCrewMember from '../components/CrewMember/FindCrewMember.vue'
 import ManageCrewMembers from '@/components/CrewMember/ManageCrewMembers.vue'
 import LoginView from '../views/LoginView.vue'
-import NavBar from '@/components/Navigation/NavBar.vue'
+import AdminNavBar from '@/components/Navigation/AdminNavBar.vue'
 import AddAvailability from '@/components/CrewMember/AddAvailability.vue'
 import CreateProfile from '@/components/CrewMember/CreateProfile.vue'
 import AdminHomeView from '@/views/AdminHomeView.vue'
